Enabled SSL for Spring Application using Self-Signed Certificate
I modified the Spring application and Dockerfile to include steps regarding setting up the Spring Application with an (Self-signed) SSL Certificate to allow the server to be accessed with `https`, which should allow for end-2-end encryption and thus prevent any hacker from using packet sniffers to get sensitive information when a user tries to log in or register an account.

Due to issues regarding Testcontainer and the inability of the Spring application to access the generated Testcontainer, I removed everything related to Testcontainers from the java files under `src/test/java/com/focust/api`. The application should still run just fine, as Testcontainer isn't going to be used until a future commit when I implement models, controllers, and repositories.