Included Refresh Tokens & Updated README.md files
For this commit, I simply included a new function for the `JWTService` class (renamed to `JwtService`) to generate refresh tokens, which will be used to allow one to not have to sign in every 5 minutes while also keeping the expiration date for access tokens short for security reasons. The refresh tokens are made to last 1 week (7 days).

I also updated the `README.md` to make it more clear on the most important dependencies for both the back-end and front-end servers, plus more general dependencies like Docker and Jenkins. Regarding the `./spring/README.md` file, I simply changed "**JUnit**" to "**JUnit 5**", as JUnit 5 is used for the tests.