Added custom conf.d directory
For this commit, I managed to figure out the issue with running the testcontainers, and it deals with a problem inside of `/etc/mysql/conf.d/my.cnf` file in said containers, where there are variables that have since then been removed as of `mysql:9.3.0` (`latest` at the time of this commit), so I made a custom `my.cnf` file that simply removes the problematic variables.

I updated the `BUILD_USING_DOCKER.md` and `Jenkinsfile` accordingly.