Changed request body for "GET /auth/refresh/"
For this commit, I decided to change the request body for the function associated with `GET /auth/refresh/` REST command from `@CookieValue(...) String refreshToken` to just `HttpServletRequest request`, and have the function extract the cookie from the cookies in said request. This is done to fix a problem regarding getting the cookie from the request to begin with.