Added code to check for /auth/register response in givenRefreshTokenCookie_whenSendingRequestForNewAccessToken_thenOkStatus
For this commit, I noticed that, for some reason, the test was unable to find the email `cookie-crumbl@focust.local` in the Testcontainer MySQL Database when evaluating the refresh token. Thus, I decided to add code to see what the response from registering the new user looked like to see if I can get some clues as to what is going on.