Added exception handling for UserNotFoundException for /auth/refresh
For this commit, I realized that I should've handled the `UserNotFoundException` thrown when making a request for a new access token, so I did just that. I also added a message that would be printed out if there was any `JWTVerificationException` or `IOException` that was caught when trying to get the email of the user from a JWT token to help me better debug the problem I am currently facing with being unable to find an already registered user.