Changed extraction of email from validated token in JwtService
For this commit, I managed to deduce that the problem was within the `getEmail()` method in `JwtService.java`. I decided to modify the extraction by first having the method check if the claim was available inside of the validated token, and then extract the email. It was through this I noticed that I was using `toString()` instead of `asString()` when getting the email from the claim. Hopefully this might fix it.