Added cookie to servletResponse in refreshAccessToken
For this commit, I managed to fix the problem of the server being unable to get the email from the cookie, but now I have another problem: there is no `jwt-refresh-token` cookie in the response for `/auth/refresh`. However, this is an easy fix as I could just add `servletResponse.addCookie(jwtRefreshTokenCookie.get())` to add the cookie back in. Hopefully, every single test will pass this time and I can merge everything done so far into the master so I can work on the front-end.