Final commit before merge
This the final commit before merging the contents of this branch to master, as every single test was able to pass and the server is able to run. However, I merely removed any code that might leak the JWT Refresh token from the main codebase as I do not want that information leaked if some hacker manages to get into the computer where the back-end is running, especially when the refresh token has a longer lifespan than a typical access token. The code should still work.