Set up the ./react directory and React project
For this commit, I have added a new directory, `./react` for the React project, which has Vite and TailwindCSS installed and uses TypeScript. For now, the application should show the text "Hello, World!".

I have added favicons, for both the light and dark mode for browsers, and modified the `Jenkinsfile` and `BUILD_USING_DOCKER.md` to include instructions to run the newly created `./react/Dockerfile` to run the Vite server, as well as the necessary networks.