Fixed front-end port error and properly assigned docker networks
For this commit, I changed `-p 5080:80` to `-p 5080:5080` as `http://localhost:80` wasn't working, and I wonder if it had something to do with that. I also realized that I accidentally added the MySQL container under the `react-spring` network instead of the back-end server, so I had to do that as well.