Added HTTPS protocol for React Vite server
For this commit, went through the process of making the React server use HTTPS for interactions, like with the back-end server, all for security reasons. The process is much more simple compared to the back-end as all I had to do is to configure the Vite server (via `vite.config.ts` file to accept two `.pem` files under the `https:` block. I also changed the port from `5080` to `5443`, as `443` is the port number for HTTPS.

I also updated the `BUILD_USING_DOCKER.md` and `BUILD_USING_JENKINS.md` accordingly.