Added '|| exit 0' to 'test -f' commands
For this commit, I realized that, while the previous commit worked, the build ended up failing to to the `test -f` command for checking if the two `.pem` files for the React front-end server. This made me think that the problem might be that, because the file doesn't exist, it returns the exit code of 1, which ends the entire Jenkins build, which is not what I want as those are only for merely removing a file if they exist.  Thus, I added `|| exit 0` at the end of each of the `test -f` to see if that would fix it.