Updated react/Dockerfile to better handle extract of certificates
For this commit, there was an error from the `react/Dockerfile` where it was unable to extract the generated `.certs` directory, as it is not under the `./react/` directory, so I modified the Dockerfile to properly get it from the *parent* directory of the directory where the Dockerfile is located. I also removed a useless command from the Dockerfile as well.