Changed port of React Vite server
For this commit, I finally managed to get all of the docker containers running, but accessing `https://localhost:443` resulted in a `SSL_ERROR_RX_RECORD_TOO_LONG`  appearing on the screen. Looking more into it, this error could be caused by mismatched port numbers, so I decided to make it easier on myself and only use the port `5443` when accessing the site, and hopefully this error would go away.